+++ server/lib/room-channel/room-channel-class.ts
@@ -352,7 +352,7 @@ class RoomChannel {
         // This can be done without waiting for the API call to finish, but we don't want to send thousands of
         // API calls at the same time. So storing data in a map, and we well launch it sequentially at the end
         prosodyRoomUpdates.set(roomJID, {
-          slow_mode_duration: channelConfigurationOptions.slowMode.defaultDuration
+          slow_mode_duration: channelConfigurationOptions.slowMode.duration
         })
 
         this.roomConfToUpdate.delete(roomJID)