+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Updating xmppjs-chat-box version.
 * Translation updates: japanese.
 * Updating slow mode duration on existing rooms when changing channel options (related to #332).
+  * This settings is no more a default duration, but the actual duration (updating labels and documentation).
 
 ## 8.3.2